Refactor: Organize test directory and fix autoloader issues
