+++ tests/bootstrap.php
@@ -23,7 +23,7 @@ define( 'AI_MIN_PHP_VERSION', '7.4' );
 define( 'AI_MIN_WP_VERSION', '6.7' );
 
 // Give access to tests_add_filter() function.
-require_once $_tests_dir . '/includes/functions.php';
+require_once $_tests_dir . '/tests/phpunit/includes/functions.php';
 
 /**
  * Manually load the plugin being tested.
@@ -34,4 +34,4 @@ function _manually_load_plugin() {
 tests_add_filter( 'muplugins_loaded', '_manually_load_plugin' );
 
 // Start up the WP testing environment.
-require $_tests_dir . '/includes/bootstrap.php';
+require $_tests_dir . '/tests/phpunit/includes/bootstrap.php';
